# frontend/src/api/generated/personnel.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DutyFlow personnel",
    "version": "0.1.0"
  },
  "paths": {
...
    "/positions": {
      "get": {
        "tags": [
          "references"
        ],
        "summary": "List Positions",
        "operationId": "list_positions_positions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PositionOut"
                  },
                  "type": "array",
                  "title": "Response List Positions Positions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "references"
        ],
        "summary": "Create Position",
        "operationId": "create_position_positions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/positions/{item_id}": {
      "put": {
        "tags": [
          "references"
        ],
        "summary": "Update Position",
        "operationId": "update_position_positions__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/person-categories": {
      "get": {
        "tags": [
          "references"
        ],
        "summary": "List Person Categories",
        "operationId": "list_person_categories_person_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonCategoryOut"
                  },
                  "type": "array",
                  "title": "Response List Person Categories Person Categories Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "references"
        ],
        "summary": "Create Person Category",
        "operationId": "create_person_category_person_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCategoryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonCategoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/person-categories/{item_id}": {
      "put": {
        "tags": [
          "references"
        ],
        "summary": "Update Person Category",
        "operationId": "update_person_category_person_categories__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCategoryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonCategoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/attribute-definitions": {
      "get": {
        "tags": [
          "references"
        ],
        "summary": "List Attribute Definitions",
        "operationId": "list_attribute_definitions_attribute_definitions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttributeDefinitionOut"
                  },
                  "type": "array",
                  "title": "Response List Attribute Definitions Attribute Definitions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "references"
        ],
        "summary": "Create Attribute Definition",
        "operationId": "create_attribute_definition_attribute_definitions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttributeDefinitionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributeDefinitionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/attribute-definitions/{item_id}": {
      "put": {
        "tags": [
          "references"
        ],
        "summary": "Update Attribute Definition",
        "operationId": "update_attribute_definition_attribute_definitions__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttributeDefinitionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributeDefinitionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exemption-reasons": {
      "get": {
        "tags": [
          "references"
        ],
        "summary": "List Exemption Reasons",
        "operationId": "list_exemption_reasons_exemption_reasons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExemptionReasonOut"
                  },
                  "type": "array",
                  "title": "Response List Exemption Reasons Exemption Reasons Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "references"
        ],
        "summary": "Create Exemption Reason",
        "operationId": "create_exemption_reason_exemption_reasons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExemptionReasonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExemptionReasonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/exemption-reasons/{item_id}": {
      "put": {
        "tags": [
          "references"
        ],
        "summary": "Update Exemption Reason",
        "operationId": "update_exemption_reason_exemption_reasons__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExemptionReasonIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExemptionReasonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttributeDefinitionIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "value_type": {
            "type": "string",
            "enum": [
              "bool",
              "int",
              "enum",
              "date",
              "string"
            ],
            "title": "Value Type"
          },
          "enum_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Options"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required",
            "default": false
          },
          "sort_order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "value_type"
        ],
        "title": "AttributeDefinitionIn"
      },
      "AttributeDefinitionOut": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "value_type": {
            "type": "string",
            "enum": [
              "bool",
              "int",
              "enum",
              "date",
              "string"
            ],
            "title": "Value Type"
          },
          "enum_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Options"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required",
            "default": false
          },
          "sort_order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "value_type",
          "id"
        ],
        "title": "AttributeDefinitionOut"
      },
...
      "ExemptionReasonIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "ExemptionReasonIn"
      },
      "ExemptionReasonOut": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "id"
        ],
        "title": "ExemptionReasonOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonCategoryIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "sort_order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "PersonCategoryIn"
      },
      "PersonCategoryOut": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "sort_order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "id"
        ],
        "title": "PersonCategoryOut"
      },
...
      "PositionIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "sort_order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PositionIn"
      },
      "PositionOut": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "sort_order": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Sort Order",
            "default": 0
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "PositionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}